+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Variable index expression is not evaluated correctly on macro calls
 - Macro calls originating from statements partially consumed by AREAD instructions might not be performed
 - The language server might crash while evaluating symbol dependencies
+- AREAD modifies macro processing stack
 
 ## [1.19.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.18.0...1.19.0) (2025-09-03)